+++ src/test/java/org/datadog/jmxfetch/SimpleTestJavaAppMBean.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 public interface SimpleTestJavaAppMBean {
     
     int getShouldBe100();
-    int getShouldBe1000();
+    Double getShouldBe1000();
     int getShouldBeCounter();
     HashMap<String, Integer> getHashmap();
